@@ -27,9 +27,9 @@
     "ra-i18n-polyglot": "^3.11.1",
     "ra-language-english": "^3.11.1",
     "ra-language-french": "^3.11.1",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-admin": "^3.19.0",
-    "react-dom": "npm:@hot-loader/react-dom",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.12.2",
     "react-hot-loader": "4.12.19",
     "react-i18next": "^11.8.5",
