@@ -35,7 +35,7 @@ import {
   useNotify,
 } from 'react-admin';
 
-import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolBar';
+import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolbar';
 import {
   validateEmail,
   validateFirstName,
