@@ -38,7 +38,7 @@ import {
   useTranslate,
 } from 'react-admin';
 
-import SimpleFormToolBar from '@components/admin/shared/toolbars/SimpleFormToolBar';
+import SimpleFormToolBar from '@components/admin/shared/toolbars/SimpleFormToolbar';
 import {
   validateCategory,
   validateSubCategory,
